requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==3.2.16
django-autoslug==1.9.8
django-tinymce==3.5.0
Pillow==9.4.0
pycodestyle==2.10.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
django-embed-video
django-ckeditor